+++ styleguide/haml_locals.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-# frozen_string_literal => true
-
 @locals = {
   "asset_hyperlink" => {
     "uri" => "//assets.ctfassets.net/vms0u05139aw/7ayjV8wUEBCqQOxaylJjxs/4b8214a762bf4e5b7b68a70bf0d1599d/test_pdf.pdf",