+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## <sub>v1.12.0</sub>
 
+* NP-978 Remove advice feedback component
+
 #### _Oct. 12, 2020_
 
 * NP-756 Advisernet related content component